requirements.txt
# Landuse Global Pipeline Dependencies

# Core scientific computing
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0
scipy>=1.7.0,<2.0.0

# Geospatial data processing
xarray>=2022.3.0
dask[complete]>=2022.1.0
netCDF4>=1.5.7
rasterio>=1.2.0
geopandas>=0.10.0
pyproj>=3.2.0
shapely>=1.8.0

# Machine learning
scikit-learn>=1.0.0,<2.0.0
tensorflow>=2.11.0,<2.16.0
scikeras>=0.12.0

# Cloud storage (optional)
google-cloud-storage>=2.0.0
gcsfs>=2022.1.0

# Optimization
scipy>=1.7.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
PyYAML>=6.0
tqdm>=4.62.0
joblib>=1.1.0
click>=8.0.0

# Logging
python-json-logger>=2.0.0

# Development dependencies (optional)
# pytest>=7.0.0
# pytest-cov>=3.0.0
# black>=22.0.0
# flake8>=4.0.0
# mypy>=0.950
